[Test] Fixes? to two issues.
The first one is changing the LatinIME package name (it gets a pretty stock name, maybe that name it's on the lists of opengapps and when it checks to remove old/obsolete apps, it takes the gboard folder and deletes it)
The second one is related to wallpaper, i'm doing some wallpaper changes, and on the overlay folder there's some stuff related to pixelwalls. i didn't include it. and there's some overlays to google wallpaper picker that's not included on all the roms. so, i left only the overlays related to google dialer.